Добавил ручки для PlacementObject и Status, рефакторинг кода.

+++ media-service/src/main/java/ru/media/accounting/service/CategoryService.java
@@ -1,6 +1,7 @@
 package ru.media.accounting.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import ru.media.accounting.dto.category.CategoryRequest;
 import ru.media.accounting.exception.ElementAlreadyExistsException;
@@ -11,15 +12,17 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 @Service
-@RequiredArgsConstructor
+@RequiredArgsConstructor(onConstructor = @__(@Lazy))
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final MediaService mediaService;
 
     /**
      * Получить категорию по названию.
      * @param title название категории. Например: "Указатель"
      * @return {@link Category} категорию.
+     * @throws NoSuchElementException если категория не найдена.
      */
     public Category findByTitle(String title) {
         return categoryRepository.findByTitle(title)
@@ -29,6 +32,7 @@ public class CategoryService {
     /**
      * Получить все категории.
      * @return список категорий.
+     * @throws NoSuchElementException если категории не найдены.
      */
     public List<Category> findAll(){
         List<Category> categoryList = categoryRepository.findAll();
@@ -42,6 +46,7 @@ public class CategoryService {
      * Создать новую категорию.
      * @param categoryRequest запрос категории.
      * @return {@link Category} категорию.
+     * @throws ElementAlreadyExistsException если категория уже существует.
      */
     public Category save(CategoryRequest categoryRequest){
         if(categoryRepository.findByTitle(categoryRequest.getTitle()).isPresent()) {
@@ -67,11 +72,16 @@ public class CategoryService {
     }
 
     /**
-     * Удалить категорию.
+     * Удалить категорию. Если категория содержится в носителе, выбрасывает исключение.
      * @param categoryRequest запрос категории.
+     * @throws ElementAlreadyExistsException если категория содержится в носителе.
      */
     public void delete(CategoryRequest categoryRequest) {
         Category category = findByTitle(categoryRequest.getTitle());
-        categoryRepository.delete(category);
+        if (mediaService.findByCategory(category).isEmpty()) {
+            categoryRepository.delete(category);
+        } else {
+            throw new ElementAlreadyExistsException("Есть носители c категорией " + categoryRequest.getTitle());
+        }
     }
 }
